% Auto-generated: do not edit by hand
\name{monacoDashEditor}

\alias{monacoDashEditor}

\title{MonacoDashEditor component}

\description{

}

\usage{
monacoDashEditor(id=NULL, className=NULL, height=NULL, language=NULL,
options=NULL, readOnly=NULL, style=NULL, theme=NULL,
value=NULL)
}

\arguments{
\item{id}{Character | lists containing elements 'id', '__extras__'.
those elements have the following types:
  - id (character; required)
  - __extras__ (list with named elements and values of type character; optional). The ID used to identify this component in Dash callbacks.
Can be a string or a dictionary with additional data attributes.}

\item{className}{Character. Additional CSS class for the container}

\item{height}{Character. Height of the editor}

\item{language}{Character. The language of the editor}

\item{options}{Named list. Additional Monaco Editor options}

\item{readOnly}{Logical. Make editor read-only}

\item{style}{Named list. Inline styles for the container}

\item{theme}{Character. Editor theme}

\item{value}{Character. Current code/content}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

